move authButton into Navigation
