+++ src/components/Header/Header.css
@@ -10,34 +10,6 @@
   box-sizing: border-box;
 }
 
-.header__button {
-  max-width: 176px;
-  height: 48px;
-  background-color: transparent;
-}
-
-.header__button:hover {
-  opacity: 0.9;
-}
-
-.header__button_white {
-  color: var(--main-bg-color);
-  border: 1px solid var(--main-bg-color);
-}
-
-.header__button_dark {
-  color: var(--main-text-color);
-  border: 1px solid var(--main-text-color);
-}
-
-.header__button-text {
-  display: block;
-  max-width: 138px;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
 .header__menu-icon:after,
 .header__menu-icon span {
   background-color: var(--main-bg-color);
@@ -52,10 +24,6 @@
 @media screen and (max-width: 768px) {
   .header {
     height: 66px;
-  }
-
-  .header__button {
-    height: 40px;
   }
 }
 
@@ -103,24 +71,5 @@
     display: grid;
     padding: 16px 0 0 0;
     margin: 0;
-  }
-
-  .header__button {
-    display: none;
-  }
-
-  .header__button_mobile {
-    width: 100%;
-    max-width: 100%;
-    grid-column: 1/3;
-    grid-row: 3;
-    display: flex;
-    justify-content: center;
-    height: 56px;
-    margin: 22px 0 24px;
-  }
-
-  .header__button-text {
-    max-width: 80%;
   }
 }
